requirements.txt
absl-py==2.3.1
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
clarabel==0.11.1
click==8.2.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cssselect2==0.8.0
cvxopt==1.3.2
cvxpy==1.7.5
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
et_xmlfile==2.0.0
executing==2.2.0
Faker==37.3.0
fancyimpute==0.7.0
fastjsonschema==2.21.1
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
flatbuffers==25.9.23
fonttools==4.56.0
fqdn==1.5.1
fsspec==2025.7.0
gast==0.7.0
google-pasta==0.2.0
grpcio==1.76.0
h11==0.14.0
h5py==3.15.1
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.33.4
idna==3.10
imbalanced-learn==0.14.0
importlib_metadata==8.6.1
iniconfig==2.3.0
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras==3.12.0
kiwisolver==1.4.8
knnimpute==0.1.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
missingno==0.5.2
mistune==3.1.3
ml_dtypes==0.5.4
MouseInfo==0.1.3
mpmath==1.3.0
namex==0.1.0
narwhals==1.41.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nose==1.3.7
notebook_shim==0.2.4
numpy==2.2.4
ollama==0.6.1
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.18.0
osqp==1.0.5
overrides==7.7.0
packaging==24.2
panda==0.3.1
pandas==2.3.0
pandoc==2.4
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
pdfkit==1.0.0
pefile==2023.2.7
pillow==11.1.0
platformdirs==4.3.7
plotly==6.1.2
pluggy==1.6.0
plumbum==1.9.0
ply==3.11
prometheus_client==0.21.1
prompt_toolkit==3.0.51
protobuf==6.33.2
psutil==7.0.0
pure_eval==0.2.3
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.12.4
pydantic_core==2.41.5
pydot==4.0.0
pydyf==0.11.0
pyee==11.1.1
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.19.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.2
PyMsgBox==1.0.9
pyparsing==3.2.3
PyPDF2==3.0.1
pyperclip==1.9.0
pyphen==0.17.2
pyppeteer==2.0.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.2
PyQt6==6.10.1
PyQt6-Qt6==6.10.1
PyQt6_sip==13.10.3
PyRect==0.2.0
PyScreeze==1.0.1
PySide6==6.10.0
PySide6_Addons==6.10.0
PySide6_Essentials==6.10.0
pytest==9.0.2
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytweening==1.2.0
pytz==2025.2
pywhatkit==5.4
pywin32==310
pywin32-ctypes==0.2.3
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.4.0
redis==6.1.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.2.0
rpds-py==0.24.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
scs==3.2.9
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==78.1.0
shiboken6==6.10.0
simsimd==6.5.12
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
spotipy==2.25.1
stack-data==0.6.3
statsmodels==0.14.4
stringzilla==4.6.0
sympy==1.13.3
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tkcalendar==1.6.1
tokenizers==0.21.2
torch==2.9.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.2
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==1.26.20
wcwidth==0.2.13
weasyprint==65.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.1.3
wheel==0.45.1
wikipedia==1.4.0
wrapt==2.0.1
zipp==3.21.0
zopfli==0.2.3.post1
